+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/user/follow/UserFollowSortFilterController.kt
@@ -1,0 +1,47 @@
+package com.thekeeperofpie.artistalleydatabase.anime.user.follow
+
+import androidx.compose.runtime.snapshotFlow
+import androidx.compose.ui.unit.dp
+import com.thekeeperofpie.artistalleydatabase.android_utils.FeatureOverrideProvider
+import com.thekeeperofpie.artistalleydatabase.anime.AnimeSettings
+import com.thekeeperofpie.artistalleydatabase.anime.R
+import com.thekeeperofpie.artistalleydatabase.anime.filter.SortFilterController
+import com.thekeeperofpie.artistalleydatabase.anime.filter.SortFilterSection
+import com.thekeeperofpie.artistalleydatabase.anime.user.UserSortOption
+import com.thekeeperofpie.artistalleydatabase.compose.filter.SortEntry
+import kotlinx.coroutines.FlowPreview
+import kotlinx.coroutines.flow.debounce
+import kotlin.time.Duration.Companion.milliseconds
+
+@OptIn(FlowPreview::class)
+class UserFollowSortFilterController(
+    settings: AnimeSettings,
+    featureOverrideProvider: FeatureOverrideProvider,
+) : SortFilterController(settings, featureOverrideProvider) {
+
+    private val sortSection = SortFilterSection.Sort(
+        enumClass = UserSortOption::class,
+        defaultEnabled = UserSortOption.ID,
+        headerTextRes = R.string.anime_user_filter_sort_label,
+    ).apply {
+        sortOptions = sortOptions.filterNot { it.value == UserSortOption.SEARCH_MATCH }
+    }
+
+    override var sections = listOf(
+        sortSection,
+        advancedSection,
+        SortFilterSection.Spacer(height = 32.dp),
+    )
+
+    fun filterParams() = snapshotFlow {
+        FilterParams(
+            sort = sortSection.sortOptions,
+            sortAscending = sortSection.sortAscending,
+        )
+    }.debounce(500.milliseconds)
+
+    data class FilterParams(
+        val sort: List<SortEntry<UserSortOption>>,
+        val sortAscending: Boolean,
+    )
+}